+++ agents/adv_sql_music_store_supervisor.py
@@ -1,5 +1,5 @@
 from agents.invoice_agent import graph as invoice_agent
-from agents.advanced_sql_agent import graph as music_agent
+from agents.adv_sql_agent import graph as music_agent
 from agents.utils import llm
 
 from typing import Annotated, NotRequired
